requirements.txt
Flask
mysql-connector-python
Flask-CORS
python-dotenv
requests
google-generativeai
